feature/error-handler-bugsnag-ignored-exceptions-resolver
* review fixes
