+++ packages/EasyErrorHandler/src/Bridge/Symfony/Builder/ApiPlatformBuilderProvider.php
@@ -35,7 +35,8 @@ final class ApiPlatformBuilderProvider implements ErrorResponseBuilderProviderIn
     public function getBuilders(): iterable
     {
         if (\class_exists(ValidationException::class)) {
-            yield new ApiPlatformErrorResponseBuilder($this->translator, $this->keys);
+            yield new ApiPlatformValidationExceptionResponseBuilder($this->translator, $this->keys);
+            yield new ApiPlatformValidationErrorResponseBuilder();
         }
     }
 }
