+++ packages/EasyErrorHandler/src/Bridge/Symfony/Builder/ApiPlatformValidationErrorResponseBuilder.php
@@ -4,16 +4,16 @@ declare(strict_types=1);
 
 namespace EonX\EasyErrorHandler\Bridge\Symfony\Builder;
 
-use ApiPlatform\Core\Bridge\Symfony\Validator\Exception\ValidationException;
 use ApiPlatform\Core\Exception\InvalidArgumentException;
+use App\Common\Exception\ErrorCodesInterface;
 use EonX\EasyErrorHandler\Builders\AbstractErrorResponseBuilder;
-use EonX\EasyErrorHandler\Interfaces\TranslatorInterface;
+use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Serializer\Exception\MissingConstructorArgumentsException;
 use Symfony\Component\Serializer\Exception\UnexpectedValueException;
 use Symfony\Component\Validator\Constraints\NotNull;
 use Throwable;
 
-final class ApiPlatformErrorResponseBuilder extends AbstractErrorResponseBuilder
+final class ApiPlatformValidationErrorResponseBuilder extends AbstractErrorResponseBuilder
 {
     private const MESSAGE_PATTERN_INVALID_DATE = '/This value is not a valid date\/time\./';
 
@@ -28,24 +28,6 @@ final class ApiPlatformErrorResponseBuilder extends AbstractErrorResponseBuilder
     ' data because its constructor requires parameter "(\w+)" to be present/';
 
     private const MESSAGE_PATTERN_TYPE_ERROR = '/The type of the "(\w+)" attribute must be "(\w+)", "(\w+)" given/';
-
-    /**
-     * @var string[]
-     */
-    private array $keys;
-
-    private TranslatorInterface $translator;
-
-    /**
-     * @param mixed[]|null $keys
-     */
-    public function __construct(TranslatorInterface $translator, ?array $keys = null, ?int $priority = null)
-    {
-        $this->translator = $translator;
-        $this->keys = $keys ?? [];
-
-        parent::__construct($priority);
-    }
 
     public static function supports(Throwable $throwable): bool
     {
@@ -74,34 +56,11 @@ final class ApiPlatformErrorResponseBuilder extends AbstractErrorResponseBuilder
      */
     public function buildData(Throwable $throwable, array $data): array
     {
-        $data['message'] = 'Validation failed.';
-
-        if ($throwable instanceof ValidationException) {
-            $violations = [];
-
-            foreach ($throwable->getConstraintViolationList() as $violation) {
-                $propertyPath = $violation->getPropertyPath();
-
-                if (isset($violations[$propertyPath]) === false) {
-                    $violations[$propertyPath] = [];
-                }
-
-                $violations[$propertyPath][] = $violation->getMessage();
-            }
-
-            $exceptionKey = $this->getKey('exception');
-            $exceptionMessageKey = $this->getKey('extended_exception_keys.message');
-
-            if (\is_array($data[$exceptionKey] ?? null)) {
-                $data[$exceptionKey][$exceptionMessageKey] = $this->translator->trans(
-                    'exceptions.entity_not_valid',
-                    []
-                );
-            }
-
-            $data[$this->getKey('message')] = $this->translator->trans('exceptions.not_valid', []);
-            $data[$this->getKey('violations')] = $violations;
+        if (self::supports($throwable) === false) {
+            return parent::buildData($throwable, $data);
         }
+
+        $data['message'] = 'Validation failed.';
 
         if ($throwable instanceof InvalidArgumentException) {
             $matches = [];
@@ -157,31 +116,10 @@ final class ApiPlatformErrorResponseBuilder extends AbstractErrorResponseBuilder
 
     public function buildStatusCode(Throwable $throwable, ?int $statusCode = null): ?int
     {
-        if ($throwable instanceof ValidationException) {
-            $statusCode = 400;
+        if (self::supports($throwable) === false) {
+            return parent::buildStatusCode($throwable, $statusCode);
         }
 
-        return parent::buildStatusCode($throwable, $statusCode);
-    }
-
-    /**
-     * @param mixed[]|null $keys
-     */
-    private function getKey(string $name, ?array $keys = null): string
-    {
-        $keys = $keys ?? $this->keys;
-        $nameParts = \explode('.', $name);
-
-        if (\count($nameParts) <= 1) {
-            return $keys[$name] ?? $name;
-        }
-
-        $firstPartOfName = \array_shift($nameParts);
-
-        if (isset($keys[$firstPartOfName]) === false) {
-            return $name;
-        }
-
-        return $this->getKey(\implode('.', $nameParts), $keys[$firstPartOfName]);
+        return Response::HTTP_BAD_REQUEST;
     }
 }
