+++ packages/EasyErrorHandler/src/Bridge/Symfony/Resources/config/bugsnag_reporter.php
@@ -5,10 +5,10 @@ declare(strict_types=1);
 namespace Symfony\Component\DependencyInjection\Loader\Configurator;
 
 use EonX\EasyErrorHandler\Bridge\BridgeConstantsInterface;
-use EonX\EasyErrorHandler\Bridge\Bugsnag\BugsnagReporterProvider;
-use EonX\EasyErrorHandler\Bridge\Bugsnag\ErrorDetailsClientConfigurator;
-use EonX\EasyErrorHandler\Bridge\Bugsnag\SeverityClientConfigurator;
-use EonX\EasyErrorHandler\Bridge\Bugsnag\UnhandledClientConfigurator;
+use EonX\EasyErrorHandler\Bridge\Bugsnag\Configurators\ErrorDetailsClientConfigurator;
+use EonX\EasyErrorHandler\Bridge\Bugsnag\Configurators\SeverityClientConfigurator;
+use EonX\EasyErrorHandler\Bridge\Bugsnag\Configurators\UnhandledClientConfigurator;
+use EonX\EasyErrorHandler\Bridge\Bugsnag\Providers\BugsnagReporterProvider;
 
 return static function (ContainerConfigurator $containerConfigurator): void {
     $services = $containerConfigurator->services();
