+++ packages/EasyErrorHandler/tests/Bridge/Bugsnag/BugsnagReporterTest.php
@@ -4,10 +4,10 @@ declare(strict_types=1);
 
 namespace EonX\EasyErrorHandler\Tests\Bridge\Bugsnag;
 
-use EonX\EasyErrorHandler\Bridge\Bugsnag\BugsnagReporter;
+use EonX\EasyErrorHandler\Bridge\Bugsnag\Interfaces\BugsnagIgnoreExceptionsResolverInterface;
+use EonX\EasyErrorHandler\Bridge\Bugsnag\Reporters\BugsnagReporter;
+use EonX\EasyErrorHandler\Bridge\Bugsnag\Resolvers\DefaultBugsnagIgnoreExceptionsResolver;
 use EonX\EasyErrorHandler\ErrorLogLevelResolver;
-use EonX\EasyErrorHandler\Interfaces\BugsnagIgnoreExceptionsResolverInterface;
-use EonX\EasyErrorHandler\Resolvers\BugsnagIgnoreExceptionsResolver;
 use EonX\EasyErrorHandler\Tests\AbstractTestCase;
 use EonX\EasyErrorHandler\Tests\Stubs\BaseExceptionStub;
 use EonX\EasyErrorHandler\Tests\Stubs\BugsnagClientStub;
@@ -78,9 +78,9 @@ final class BugsnagReporterTest extends AbstractTestCase
         ?array $ignoredExceptions = null
     ): void {
         $stub = new BugsnagClientStub();
-        $reporter = new BugsnagReporter(
+        $reporter = new \EonX\EasyErrorHandler\Bridge\Bugsnag\Reporters\BugsnagReporter(
             $stub,
-            new BugsnagIgnoreExceptionsResolver(),
+            new DefaultBugsnagIgnoreExceptionsResolver(),
             new ErrorLogLevelResolver(),
             $threshold,
             $ignoredExceptions
